Merge master branch at commit 9b0773cbd5ae ("libutils: fix memset(),
memcpy(), memmove() with -O3") which is just before a LTC patches. This
will make sure the branch is buildable before cherry-picking the LTC
patches.

+++ core/include/kernel/tee_ta_manager.h
@@ -105,7 +105,7 @@ struct tee_ta_session {
 	uint32_t ref_count;	/* reference counter */
 	struct condvar refc_cv;	/* CV used to wait for ref_count to be 0 */
 	struct condvar lock_cv;	/* CV used to wait for lock */
-	int lock_thread;	/* Id of thread holding the lock */
+	short int lock_thread;	/* Id of thread holding the lock */
 	bool unlink;		/* True if session is to be unlinked */
 #if defined(CFG_TA_GPROF_SUPPORT)
 	struct sample_buf *sbuf; /* Profiling data (PC sampling) */
